feat(设置): ✨ 增加'自动删除加载失败的伪代码文件'选项 feat(设置): ✨ 增加'自动删除加载失败的插件'选项 refactor(日志活动): ♻️ 重写日志记录活动中的日志获取逻辑

+++ app/src/main/java/com/kagg886/seiko/fragment/PluginFragment.java
@@ -13,6 +13,7 @@ import android.widget.AdapterView;
 import android.widget.ListView;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
+import androidx.preference.PreferenceManager;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.textfield.TextInputLayout;
@@ -40,7 +41,6 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
     private PluginAdapter adapter;
     private AlertDialog dialog;
 
-    private final MainActivity avt;
     private final Handler mHandler = new Handler(Looper.getMainLooper()) {
         @Override
         public void handleMessage(@NonNull Message msg) {
@@ -49,25 +49,20 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
                     PluginFragment.this.dialog.show();
                     break;
                 case 0:
-                    SnackBroadCast.sendBroadCast(avt, "下载完成");
+                    SnackBroadCast.sendBroadCast(getContext(), "下载完成");
                     PluginFragment.this.dialog.dismiss();
                     break;
                 case 1:
-                    SnackBroadCast.sendBroadCast(avt, "发生异常:" + ((Throwable) msg.getData().getSerializable("exception")).getMessage());
+                    SnackBroadCast.sendBroadCast(getContext(), "发生异常:" + ((Throwable) msg.getData().getSerializable("exception")).getMessage());
                     PluginFragment.this.dialog.dismiss();
                     break;
             }
         }
     };
 
-    public PluginFragment(MainActivity avt) {
-        super();
-        this.avt = avt;
-    }
-
     @Override
     public void onClick(View v) {
-        AlertDialog dialog = new AlertDialog.Builder(avt)
+        AlertDialog dialog = new AlertDialog.Builder(getActivity())
                 .setTitle("您要...").setItems(new String[]{"从网络导入"}, (dialog1, which) -> {
                     if (which == 0) {
                         importPluginDialog().show();
@@ -77,7 +72,7 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
     }
 
     public void downloadPlugin(MainActivity avt, String url) {
-        PluginFragment.this.dialog = new AlertDialog.Builder(avt)
+        PluginFragment.this.dialog = new AlertDialog.Builder(((MainActivity) getActivity()))
                 .setTitle("下载中...")
                 .setCancelable(false)
                 .setMessage("请稍等片刻...")
@@ -101,18 +96,20 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
                 File file = null;
                 try {
                     byte[] data = response.body().source().readByteArray();
-                    File pluginDir = avt.getExternalFilesDir("plugin");
+                    File pluginDir = ((MainActivity) getActivity()).getExternalFilesDir("plugin");
                     String fileName = UUID.randomUUID().toString().replace("-", "") + ".apk";
                     file = Paths.get(pluginDir.getAbsolutePath(), fileName).toFile();
                     IOUtil.writeByteToFile(file.getAbsolutePath(), data);
-                    avt.runOnUiThread(() -> {
+                    ((MainActivity) getActivity()).runOnUiThread(() -> {
                         BotRunnerService.INSTANCE.getSeikoPluginList().refresh();
                         adapter.notifyDataSetChanged();
                     });
                 } catch (Exception e) {
                     // 下载失败，删除文件
                     if (file != null && file.exists()) {
-                        file.delete();
+                        if (PreferenceManager.getDefaultSharedPreferences(getActivity()).getBoolean("badPluginAutoDel", true)) {
+                            file.delete();
+                        }
                     }
                     notifyError(e);
                     return;
@@ -120,36 +117,6 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
                 mHandler.sendEmptyMessage(0);
             }
         });
-//        IOUtil.asyncHttp(Jsoup.connect(url), new IOUtil.Response() {
-//            @Override
-//            public void onSuccess(byte[] data) {
-//                File file = null;
-//                try {
-//                    File pluginDir = avt.getExternalFilesDir("plugin");
-//                    String fileName = UUID.randomUUID().toString().replace("-", "") + ".apk";
-//                    file = Paths.get(pluginDir.getAbsolutePath(), fileName).toFile();
-//                    IOUtil.writeByteToFile(file.getAbsolutePath(), data);
-//                    avt.runOnUiThread(() -> {
-//                        BotRunnerService.INSTANCE.getSeikoPluginList().refresh();
-//                        adapter.notifyDataSetChanged();
-//                    });
-//                } catch (Exception e) {
-//                    // 下载失败，删除文件
-//                    if(file != null && file.exists()) {
-//                        file.delete();
-//                    }
-//                    notifyError(e);
-//                    return;
-//                }
-//                mHandler.sendEmptyMessage(0);
-//            }
-//
-//            @Override
-//            public void onFailed(IOException e) {
-//                e.printStackTrace();
-//                notifyError(e);
-//            }
-//        });
     }
 
     public void notifyError(Throwable e) {
@@ -166,12 +133,12 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
         View v = inflater.inflate(R.layout.fragment_plugin, container, false);
 
         listView = v.findViewById(R.id.fragment_plugin_list);
-        adapter = new PluginAdapter(avt);
+        adapter = new PluginAdapter(((MainActivity) getActivity()));
         listView.setAdapter(adapter);
         listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
-                androidx.appcompat.app.AlertDialog.Builder builder = new androidx.appcompat.app.AlertDialog.Builder(avt);
+                androidx.appcompat.app.AlertDialog.Builder builder = new androidx.appcompat.app.AlertDialog.Builder(((MainActivity) getActivity()));
                 SeikoPlugin desc = BotRunnerService.INSTANCE.getSeikoPluginList().get(position);
                 builder.setTitle("操作:" + desc.getDescription().getName());
                 builder.setItems(new String[]{"删除插件"}, new DialogInterface.OnClickListener() {
@@ -179,13 +146,13 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
                     public void onClick(DialogInterface dialog, int which) {
                         if (which == 0) {
                             if (desc.getFile() == null) {
-                                SnackBroadCast.sendBroadCast(avt, "内置插件不可删除");
+                                SnackBroadCast.sendBroadCast(((MainActivity) getActivity()), "内置插件不可删除");
                                 return;
                             }
                             desc.getFile().delete();
                             BotRunnerService.INSTANCE.getSeikoPluginList().remove(desc);
                             adapter.notifyDataSetChanged();
-                            SnackBroadCast.sendBroadCast(avt, "删除成功!");
+                            SnackBroadCast.sendBroadCast(((MainActivity) getActivity()), "删除成功!");
                         }
                     }
                 });
@@ -202,17 +169,17 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
     }
 
     public AlertDialog importPluginDialog() {
-        AlertDialog.Builder builder = new AlertDialog.Builder(avt);
-        View view = LayoutInflater.from(avt).inflate(R.layout.dialog_import_plugin, null);
+        AlertDialog.Builder builder = new AlertDialog.Builder(((MainActivity) getActivity()));
+        View view = LayoutInflater.from(((MainActivity) getActivity())).inflate(R.layout.dialog_import_plugin, null);
         builder.setView(view);
         builder.setPositiveButton("确定", (dialog, which) -> {
             TextInputLayout importPluginUrl = view.findViewById(R.id.dialog_importPluginUrl);
             String url = importPluginUrl.getEditText().getText().toString();
-            if(url.trim().length() == 0 || !(url.startsWith("http://") || url.startsWith("https://"))) {
-                SnackBroadCast.sendBroadCast(avt, "请正确填写链接");
+            if (url.trim().length() == 0 || !(url.startsWith("http://") || url.startsWith("https://"))) {
+                SnackBroadCast.sendBroadCast(((MainActivity) getActivity()), "请正确填写链接");
                 return;
             }
-            downloadPlugin(avt, url);
+            downloadPlugin(((MainActivity) getActivity()), url);
         });
         return builder.create();
     }
@@ -221,7 +188,7 @@ public class PluginFragment extends Fragment implements View.OnClickListener, Sw
     public void onRefresh() {
         //TODO 此处刷新一次则伪代码会刷新两次，知道原因但是不知道怎么解，求助
         if (BotRunnerService.INSTANCE == null) {
-            SnackBroadCast.sendBroadCast(avt, "服务启动中，请稍等片刻");
+            SnackBroadCast.sendBroadCast(((MainActivity) getActivity()), "服务启动中，请稍等片刻");
             layout.setRefreshing(false);
             return;
         }
